+++ src/modules/auth/pages/sign-up-or-login-page/components/sign-up-form/sign-up-form.component.html
