+++ src/modules/auth/pages/sign-up-or-login-page/components/sign-up-form/sign-up-form.component.ts
@@ -2,7 +2,7 @@ import { NgClass, NgFor, NgIf } from '@angular/common';
 import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { FormGroup, FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
-import { SignUpOrLoginErrorForm } from 'src/auth/types';
+import { SignUpOrLoginErrorForm } from 'src/modules/auth/types';
 import { ToastService } from 'src/shared/services/toast.service';
 import { FormErrorMessageComponent } from 'src/shared/ui/components/form-error-message/form-error-message.component';
