+++ src/modules/auth/pages/sign-up-or-login-page/sign-up-or-login-page.component.spec.ts
