+++ src/modules/home/pages/content-wall-page/content-wall-page.component.ts
