+++ src/modules/home/pages/content-wall-page/content-wall-page.module.ts
@@ -1,10 +1,8 @@
 import { NgModule } from '@angular/core';
 import { ContentWallPageComponent } from './content-wall-page.component';
 
-
-
 @NgModule({
   declarations: [ContentWallPageComponent],
   imports: []
 })
-export class ContentWallPageModule { }
+export class ContentWallPageModule {}
